+++ src/runtime/GlobalObject.h
@@ -48,196 +48,14 @@ public:
     friend class GlobalEnvironmentRecord;
     friend class IdentifierNode;
 
-    explicit GlobalObject(ExecutionState& state)
-        : Object(state, ESCARGOT_OBJECT_BUILTIN_PROPERTY_NUMBER, Object::__ForGlobalBuiltin__)
-        , m_context(state.context())
-        , m_object(nullptr)
-        , m_objectPrototypeToString(nullptr)
-        , m_objectCreate(nullptr)
-        , m_objectFreeze(nullptr)
-        , m_function(nullptr)
-        , m_functionPrototype(nullptr)
-        , m_functionApply(nullptr)
-        , m_iteratorPrototype(nullptr)
-        , m_error(nullptr)
-        , m_errorPrototype(nullptr)
-        , m_referenceError(nullptr)
-        , m_referenceErrorPrototype(nullptr)
-        , m_typeError(nullptr)
-        , m_typeErrorPrototype(nullptr)
-        , m_rangeError(nullptr)
-        , m_rangeErrorPrototype(nullptr)
-        , m_syntaxError(nullptr)
-        , m_syntaxErrorPrototype(nullptr)
-        , m_uriError(nullptr)
-        , m_uriErrorPrototype(nullptr)
-        , m_evalError(nullptr)
-        , m_evalErrorPrototype(nullptr)
-        , m_string(nullptr)
-        , m_stringPrototype(nullptr)
-        , m_stringIteratorPrototype(nullptr)
-        , m_number(nullptr)
-        , m_numberPrototype(nullptr)
-        , m_symbol(nullptr)
-        , m_symbolPrototype(nullptr)
-        , m_array(nullptr)
-        , m_arrayPrototype(nullptr)
-        , m_arrayIteratorPrototype(nullptr)
-        , m_arrayPrototypeValues(nullptr)
-        , m_boolean(nullptr)
-        , m_booleanPrototype(nullptr)
-        , m_date(nullptr)
-        , m_datePrototype(nullptr)
-        , m_regexp(nullptr)
-        , m_regexpPrototype(nullptr)
-        , m_regexpSplitMethod(nullptr)
-        , m_regexpReplaceMethod(nullptr)
-        , m_regexpExecMethod(nullptr)
-        , m_math(nullptr)
-        , m_eval(nullptr)
-        , m_throwTypeError(nullptr)
-        , m_throwerGetterSetterData(nullptr)
-        , m_stringProxyObject(nullptr)
-        , m_numberProxyObject(nullptr)
-        , m_booleanProxyObject(nullptr)
-        , m_symbolProxyObject(nullptr)
-        , m_json(nullptr)
-        , m_jsonStringify(nullptr)
-        , m_jsonParse(nullptr)
-#if defined(ENABLE_ICU) && defined(ENABLE_INTL)
-        , m_intl(nullptr)
-        , m_intlCollator(nullptr)
-        , m_intlDateTimeFormat(nullptr)
-        , m_intlNumberFormat(nullptr)
-#endif
-        , m_promise(nullptr)
-        , m_promisePrototype(nullptr)
-        , m_proxy(nullptr)
-        , m_reflect(nullptr)
-        , m_arrayBuffer(nullptr)
-        , m_arrayBufferPrototype(nullptr)
-        , m_dataView(nullptr)
-        , m_dataViewPrototype(nullptr)
-        , m_typedArray(nullptr)
-        , m_typedArrayPrototype(nullptr)
-        , m_int8Array(nullptr)
-        , m_int8ArrayPrototype(nullptr)
-        , m_uint8Array(nullptr)
-        , m_uint8ArrayPrototype(nullptr)
-        , m_uint8ClampedArray(nullptr)
-        , m_uint8ClampedArrayPrototype(nullptr)
-        , m_int16Array(nullptr)
-        , m_int16ArrayPrototype(nullptr)
-        , m_uint16Array(nullptr)
-        , m_uint16ArrayPrototype(nullptr)
-        , m_int32Array(nullptr)
-        , m_int32ArrayPrototype(nullptr)
-        , m_uint32Array(nullptr)
-        , m_uint32ArrayPrototype(nullptr)
-        , m_float32Array(nullptr)
-        , m_float32ArrayPrototype(nullptr)
-        , m_float64Array(nullptr)
-        , m_float64ArrayPrototype(nullptr)
-        , m_map(nullptr)
-        , m_mapPrototype(nullptr)
-        , m_mapIteratorPrototype(nullptr)
-        , m_set(nullptr)
-        , m_setPrototype(nullptr)
-        , m_setIteratorPrototype(nullptr)
-        , m_weakMap(nullptr)
-        , m_weakMapPrototype(nullptr)
-        , m_weakSet(nullptr)
-        , m_weakSetPrototype(nullptr)
-        , m_generatorFunction(nullptr)
-        , m_generator(nullptr)
-        , m_generatorPrototype(nullptr)
-        , m_asyncFunction(nullptr)
-        , m_asyncFunctionPrototype(nullptr)
-        , m_asyncIteratorPrototype(nullptr)
-        , m_asyncFromSyncIteratorPrototype(nullptr)
-        , m_asyncGenerator(nullptr)
-        , m_asyncGeneratorPrototype(nullptr)
-        , m_asyncGeneratorFunction(nullptr)
-    {
-        m_objectPrototype = Object::createBuiltinObjectPrototype(state);
-
-        Object::setPrototype(state, m_objectPrototype);
-        Object::setGlobalIntrinsicObject(state);
-    }
+    explicit GlobalObject(ExecutionState& state);
 
     virtual bool isGlobalObject() const override
     {
         return true;
     }
 
-    void installBuiltins(ExecutionState& state)
-    {
-        installFunction(state);
-        installObject(state);
-        installIterator(state);
-        installError(state);
-        installSymbol(state);
-        installString(state);
-        installNumber(state);
-        installBoolean(state);
-        installArray(state);
-        installMath(state);
-        installDate(state);
-        installRegExp(state);
-        installJSON(state);
-#if defined(ENABLE_ICU) && defined(ENABLE_INTL)
-        installIntl(state);
-#endif
-        installPromise(state);
-        installProxy(state);
-        installReflect(state);
-        installDataView(state);
-        installTypedArray(state);
-        installMap(state);
-        installSet(state);
-        installWeakMap(state);
-        installWeakSet(state);
-        installGenerator(state);
-        installAsyncFunction(state);
-        installAsyncIterator(state);
-        installAsyncFromSyncIterator(state);
-        installAsyncGeneratorFunction(state);
-        installOthers(state);
-    }
-
-    void installFunction(ExecutionState& state);
-    void installObject(ExecutionState& state);
-    void installError(ExecutionState& state);
-    void installSymbol(ExecutionState& state);
-    void installString(ExecutionState& state);
-    void installNumber(ExecutionState& state);
-    void installBoolean(ExecutionState& state);
-    void installArray(ExecutionState& state);
-    void installMath(ExecutionState& state);
-    void installDate(ExecutionState& state);
-    void installRegExp(ExecutionState& state);
-    void installJSON(ExecutionState& state);
-#if defined(ENABLE_ICU) && defined(ENABLE_INTL)
-    void installIntl(ExecutionState& state);
-#endif
-    void installPromise(ExecutionState& state);
-    void installProxy(ExecutionState& state);
-    void installReflect(ExecutionState& state);
-    void installDataView(ExecutionState& state);
-    void installTypedArray(ExecutionState& state);
-    template <typename TA, int elementSize, typename TypeAdaptor>
-    FunctionObject* installTypedArray(ExecutionState& state, AtomicString taName, Object** proto, FunctionObject* typedArrayFunction);
-    void installIterator(ExecutionState& state);
-    void installMap(ExecutionState& state);
-    void installSet(ExecutionState& state);
-    void installWeakMap(ExecutionState& state);
-    void installWeakSet(ExecutionState& state);
-    void installGenerator(ExecutionState& state);
-    void installAsyncFunction(ExecutionState& state);
-    void installAsyncIterator(ExecutionState& state);
-    void installAsyncFromSyncIterator(ExecutionState& state);
-    void installAsyncGeneratorFunction(ExecutionState& state);
-    void installOthers(ExecutionState& state);
+    void installBuiltins(ExecutionState& state);
 
     Value eval(ExecutionState& state, const Value& arg);
     Value evalLocal(ExecutionState& state, const Value& arg, Value thisValue, InterpretedCodeBlock* parentCodeBlock, bool inWithOperation); // we get isInWithOperation as parameter because this affects bytecode
@@ -870,6 +688,40 @@ private:
     Object* m_asyncGenerator; // %AsyncGenerator%
     Object* m_asyncGeneratorPrototype; // %AsyncGeneratorPrototype%
     FunctionObject* m_asyncGeneratorFunction; // %AsyncGeneratorFunction%
+
+    void installFunction(ExecutionState& state);
+    void installObject(ExecutionState& state);
+    void installError(ExecutionState& state);
+    void installSymbol(ExecutionState& state);
+    void installString(ExecutionState& state);
+    void installNumber(ExecutionState& state);
+    void installBoolean(ExecutionState& state);
+    void installArray(ExecutionState& state);
+    void installMath(ExecutionState& state);
+    void installDate(ExecutionState& state);
+    void installRegExp(ExecutionState& state);
+    void installJSON(ExecutionState& state);
+#if defined(ENABLE_ICU) && defined(ENABLE_INTL)
+    void installIntl(ExecutionState& state);
+#endif
+    void installPromise(ExecutionState& state);
+    void installProxy(ExecutionState& state);
+    void installReflect(ExecutionState& state);
+    void installDataView(ExecutionState& state);
+    void installTypedArray(ExecutionState& state);
+    template <typename TA, int elementSize, typename TypeAdaptor>
+    FunctionObject* installTypedArray(ExecutionState& state, AtomicString taName, Object** proto, FunctionObject* typedArrayFunction);
+    void installIterator(ExecutionState& state);
+    void installMap(ExecutionState& state);
+    void installSet(ExecutionState& state);
+    void installWeakMap(ExecutionState& state);
+    void installWeakSet(ExecutionState& state);
+    void installGenerator(ExecutionState& state);
+    void installAsyncFunction(ExecutionState& state);
+    void installAsyncIterator(ExecutionState& state);
+    void installAsyncFromSyncIterator(ExecutionState& state);
+    void installAsyncGeneratorFunction(ExecutionState& state);
+    void installOthers(ExecutionState& state);
 };
 }
