+++ src/runtime/PointerValue.cpp
@@ -24,6 +24,11 @@
 
 namespace Escargot {
 
+size_t PointerValue::g_arrayObjectTag;
+size_t PointerValue::g_arrayPrototypeObjectTag;
+size_t PointerValue::g_objectRareDataTag;
+size_t PointerValue::g_doubleInSmallValueTag;
+
 Value PointerValue::call(ExecutionState& state, const Value& thisValue, const size_t argc, NULLABLE Value* argv)
 {
     ASSERT(!isCallable());
